build: Change compilation options for the PolarisEndpointResolver.
