+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.qcloud.cos</groupId>
     <artifactId>hadoop-cos</artifactId>
-    <version>8.3.10</version>
+    <version>8.3.11</version>
     <packaging>jar</packaging>
 
     <name>Apache Hadoop Tencent Cloud COS Support</name>
@@ -59,7 +59,7 @@
         <maven.compiler.source>1.7</maven.compiler.source>
         <maven.compiler.target>1.7</maven.compiler.target>
         <hadoop.version>2.8.5</hadoop.version>
-        <cos_api.version>5.6.137.2</cos_api.version>
+        <cos_api.version>5.6.228</cos_api.version>
         <google.guava.version>24.1.1-jre</google.guava.version>
         <commons_lang3.version>3.1</commons_lang3.version>
         <junit.version>4.8</junit.version>
@@ -169,6 +169,7 @@
                         <configuration>
                             <excludes>
                                 <exclude>org/apache/hadoop/fs/cosn/TencentCloudL5EndpointResolverImpl.java</exclude>
+                                <exclude>org/apache/hadoop/fs/cosn/TencentPolarisEndpointResolverImpl.java</exclude>
                             </excludes>
                         </configuration>
                     </plugin>
